Working WebApp
This is the complete working example from which Moodle tutorials are based. Students may use this to have full functionality of their project if they did not complete the project.
Note: if an I2C device is not connected, then the I2C device should be removed from the code as there is no exception handling around not connecting to a device.